# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TIDE-Mem",
    "description": "Evidence-only memory API for the Agent Memory Challenge 2026.",
    "version": "0.1.0-amc2026"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory/add": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Add Memory",
        "operationId": "add_memory_v1_memory_add_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory/search": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Search Memory",
        "operationId": "search_memory_v1_memory_search_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Request Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MemoryMessage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Messages"
          },
          "user_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "User Id"
          },
          "session_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "messages",
          "user_id",
          "session_id"
        ],
        "title": "AddRequest"
      },
      "AddResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "user_id",
          "session_id"
        ],
        "title": "AddResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "default": "ok"
          },
          "system": {
            "type": "string",
            "title": "System"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "llm_mode": {
            "type": "string",
            "title": "Llm Mode"
          }
        },
        "type": "object",
        "required": [
          "system",
          "version",
          "llm_model",
          "llm_mode"
        ],
        "title": "HealthResponse"
      },
      "MemoryMessage": {
        "properties": {
          "role": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Role"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "MemoryMessage"
      },
      "SearchItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "SearchItem"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "user_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "User Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Top K"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "user_id",
          "top_k"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SearchItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}